# feat: rework child rendering to use class