@43081j 2024-04-01
Basically we spend a lot of time waiting for garbage collection each time we finish a `renderChild`

Mostly because we discard the buffer and the promise we were waiting for. 

By itself, that isn't much. But when you have thousands of promises in memory, the clean up of them takes a while

I feel like there's a solution that will just click one day in my head 😅 something to do with having one big queue and sorting it at the end before flushing it